Ensure the observation is 'zero-cost' for non-users.
Profiling the code, I found about a ~1.5k steps per second hit on just
including the existing code - even with observations for tty_* turned
off. This must be due to the terminal emulator that goes ahead
regardless. The emulator is now not run unless observations are desired,
resulting in the reclaiming of that performance.

The profiling script for these will be included in another PR, which I will indicate
on GH.