Reduce performance costs due to screen descriptions.  (#131)
There are two performance improvements.
* The first is only storing screen_descriptions if the observation key
is present. This increases performance by about  ~2-3k sps on average,
for those environments that don't use the observation key
* The second is changing from a manual copy to strncpy over c_strings.
This seems to add another ~3k to the performance of the screen_descriptions
observation, as we can now use the very fast memcpy and strncpy for our copy operations.