Cherry-pick upstream fix e7e07cb0.
This change by PatR was in response to our PR at
https://github.com/NetHack/NetHack/pull/411 and fixes an issue
with "(null)" on MacOS. Original commit message below.

Purpose of cherry-picking this is to get this MacOS fix and also
diverge less from upstream NetHack.

----

pull request #411 - freeing termcap hilite entries

(strings to switch color) for ANSI_DEFAULT.  Instead of lumping
more conditional code into tty_shutdown() I put the new code
into a separate routine and also pulled the existing setup code
out of tty_startup() into a separate routine too.

It will be a miracle if this doesn't break anything due to the
crazy amount of convoluted conditionals present in termcap.c.

On the other hand, I found and fixed a bug while trying to test.
The ANSI_DEFAULT hilites for Gray and No_Color were null instead
of an empty string.  MS-DOS stdio apparently fixes that up, but
on OSX (after #undef UNIX and TERMLIB and TERMINFO and #define
ANSI_DEFAULT in termcap.c) I started seeing instances of "(null)"
on the map (OSX stdio does a different fix up for Null pointers)
as soon as I enabled 'color'.  It was an attempt to set No_Color.

Closes #411