agent: Add RLLib Agent
Add capability to train agents on Nethack using RLLib. Currently a lot
of the configuration values are left as default with no command-line
options of changing them; I'll adjust that later.

I used an outdated model definintion (i.e. the one in the master branch
of the nle repository, not the new one found in
nle/agent/models/base.py. This could be changed but will require
additional adjustment of configuration.

This can be run (from the root of the repository) with
`python nle/agents/rllib_agent/run.py`, optionally passing more
configuration if necessary.

It's setup to use wandb if you specify that in the command line.

I think it would make sense (once we have all the agents implemented) to
have a shared configuration framework - for example if everything used
hydra and omegaconf, then it will be easy to switch between frameworks
for the parts that are framework-agnostic (model configuration &
environment configuration). I haven't implemented that here at all so
far, but it wouldn't be too hard to do.