Add "score" to internal observation.
What exactly NH means by score is a bit context-dependent. The value
u.urexp factors into botl_score() in botl.c (if enabled) as well as
into the computation in really_done() in end.c. In many cases, the
result of botl_score() will be what really_done() reports as the
end-of-game score. In many other cases, it won't: E.g. after dying
(not quitting, escaping, ascending or bug-related exits), score for
acquired gold is reduced by 10%. Also if escaped or ascended, valuable
gems, artefacts etc will give extra score. The computation in end.c
adds to u.urexp, which will reflect the end-of-game score for a few
briefs "steps" at the very end of the game (but not before).