Default savedir to None (and remove unused archivefile)
savedir is a feature that will save ttyrecs of runs - this can be unexpected when training a model and may write a *huge* amount of unwanted files.

This changes the default to *not* save ttyrecs.

If you want ttyrecs to be saved, please specify savedir when instantiating the env from now on.

`archivefile` is a parameter that may have been used at some point, but is currently unused, so let's remove it.