Render frames even before --start.
As the comment says, could also only start rendering from the last
CLRSCREEN before FLAGS.start, which would be potentially faster.
The logic to do that requires introducing a few extra modes and
does not seem worth it at this point. The logic here is subtle
enough as it is.