Fixing up tests for converter, db and dataset.
Add a new integration test testing the construction of the dataset from
the altorg and nle_data directories. These altorg games have been checked to see if they are
consecutive episodes, and the match up is correct.

In this mock directory we include extra mock ttyrecs that should not be
included in the dataset. Readmore about the mock dataset in altorg/about.txt

Also: change dataset to default shuffle games.