Add support for v3 ttyrecs in the dataset.
Add support for v3 ttyrecs in the converter, and feed through to the
dataset. Also add support to the read_tty script.  Note the converter
now writes a score when the frame comes up, but only "flushes" the main
frame when the action comes in.

We test with ttyrec with an example of gaining core from picking up gold and
killing monsters. We expect the action at index t to correspond to the
state and reward just passed.