Add the filepath of the ttyrec if it exists to the xlogfile.
This will make it easier for users to find which ttyrecs correspond to
which metadata.  Xlogfiles are only written to with natural endings; not
guaranteed if you call "reset" before "done" is ready. This means if
there is a panic and drop a game, you will not get alignment between
your ttyrec reset number and xlogfile row. This adds an explicit link.

NB: This only works because in nle (vs typical NetHack), we play the
whole game in _one_ ttyrec.