Fixup a bug in DEC rendering.
Various commands in libtmt reset the attributes of the charact
(`attrs`). This is desirable for color (etc), but not for whether we
have DEC graphics or not. This meant if we have coloured DEC rendering,
rendering would fail.

This change ensures the permanance of the DEC flag in attrs, and updates
the test file to include this example.